Fixes problems with repeat Puppet runs.
This commit fixes the issue(s) relating to parts of the Puppet script
being re-executed repeatedly - in particular it was re-starting oracle-shm
and oracle-xe unnecessarily (after the inital run, both should already
be set up as services).

Modifies oracle-shm script to properly give status - this allows us
to put conditions on Exec["oracle-shm"] such that we can prevent executing
the command if the service is already running.

Adds requires for .rsp and sysctl files so that they're there when
we want to oracle-xe config.

Moves files from /tmp into /files so that
we don't keep having to re-provision them.